+++ CHANGELOG.md
@@ -6,6 +6,10 @@ All notable changes to this project will be documented in this file.
 
 - Describe upcoming changes here.
 
+## [0.1.2] - 2025-12-17
+- Default `INIT_SYSTEM` is now `openrc`.
+- When using OpenRC, the installer now sets up an OpenRC-managed zram swap service (`/etc/init.d/zram` + `/etc/conf.d/zram`).
+
 ## [0.1.1] - 2025-12-16
 - Semver migration (VERSION/CI alignment)